extract 'symbol' method to own module
